WIP: Rearrange function.rs into a full module; Still need to update the restore procedure
