+++ executor/read/step_executor.rs
@@ -8,7 +8,10 @@ use std::sync::Arc;
 
 use compiler::{
     executable::{
-        function::{ExecutableFunction, ExecutableFunctionRegistry, ExecutableReturn, FunctionTablingType},
+        function::{
+            executable::{ExecutableFunction, ExecutableReturn},
+            ExecutableFunctionRegistry, FunctionTablingType,
+        },
         match_::planner::match_executable::{ExecutionStep, MatchExecutable},
         next_executable_id,
         pipeline::ExecutableStage,
@@ -146,7 +149,7 @@ pub(crate) fn create_executors_for_match(
                 let _step_profile = stage_profile.extend_or_get(index, || format!("{}", function_call));
 
                 let function = function_registry.get(&function_call.function_id).unwrap();
-                if let FunctionTablingType::Tabled(_) = function.is_tabled  {
+                if let FunctionTablingType::Tabled(_) = function.is_tabled {
                     let executor = TabledCallExecutor::new(
                         function_call.function_id.clone(),
                         function_call.arguments.clone(),
